+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## WIP
+
+### Fixed
+
+- Fix Scroll component on iOS 12
+
 ## 3.14.1 (May 3, 2022)
 
 ### Added